requirements.txt
asgiref==3.2.3
Django==3.0.3
django-debug-toolbar==2.2
django-environ==0.4.5
django-filter==2.2.0
django-paranoid==0.8
django-resized==0.3.9
djangorestframework==3.11.0
djangorestframework-camel-case==1.1.2
isort==4.3.21
Markdown==3.2.1
Pillow==7.0.0
psycopg2==2.8.4
pytz==2019.3
sqlparse==0.3.0
